Central Command&ert Ships (#1443)
* Removing outdated or unused items

These ships were formalyl only really used by me likewise have minor inperfection or don't correctly work on start. So these are getting removed given other folders have been added and this one being created for ERT & CC. So just consolidating this.

* New CC & ERT Admin spawns

New ships and installations which admins can spawn to use for ingame events or as a response to ungoing things in the round.

* New CC

Entirely rebuilt. Tested if EVAC or EVAC pods dock. They appear to do so, the station is also functional.

* Fixed invalids

Thanks to dunrab. He fixed my invalids and was able to get me a media image. Thanks!

* Spawn issue + Additons

I had saved the things as mapping instead of grid, i had fixed them. Likewise made some additonal such as the ERT floating and CC relay satlite.

* Removed Decals

Was requested to remove some decals at the EVAC area.

* Incorrect formatting

This was removed due to it being incorrect formatting likewise an accident copy of the other file.

* Rename requested

This is being removed to make it more clear this a central command cargo shuttle.

* Rename of cargo shuttle

The new renamed cargo shuttle to help identify it better.

* Requested changes

These files were changed at the request of the review dunrab, in the way they had requested which was listed on the discord.

* Fixed minor errors

made some corrections at Dunrab's request.